requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-crontab==0.7.1
django-filter==2.4.0
djangorestframework==3.12.2
EasyProcess==0.3
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
mysqlclient==1.4.6
pylint==2.6.0
pytz==2020.4
PyVirtualDisplay==1.3.2
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
